Add the parlib package for akaros
Make akaros builds always include a go package called 'parlib' that
includes an import "C" directive.  By forcing this, we are able to
ensure that our final executables are linked with our cross compiler and
not 8l/6l, and we have access to all of the c code contained in our
cross compiler's shared libraries -- most notably libparlib. The plan is
to slowly create go wrappers around the functionality contained in
parlib and export it to the rest of the go library without having to
reimplement it here.