Go pops into main.main()!
Nothing really works yet because I ahven't ported the pkg/syscall stuff
yet, but we pope into main and I'm able to make goc style calls to
in there to make sure it makes it in and is executing properly.  We also
appear to be exiting cleanly, which is nice.

Next step is to port the syscall and os interfaces.