Port of the syscall pkg to Akaros
This is not yet a full port in that I've only ported over a few syscalls
to basically get "Hello World" to work and left the rest of the setup
thinking it's still operating on Linux.

That said, I've interposed at the proper place to make any syscall made
through the syscall package at least attempt to make them properly
through Akaros.  All Linux non-ported system calls (i.e. those from
Linux) are still defined, they are just set to 0, so if they ever get
called the Kernel will tell us about it.

I also worked things into the GO build system in the "right" way, in
that I'm using their mkall.sh script to parse our cross compiler headers
to automatically generate alot of the the Go code that the rest of the
packages use to interface with our system calls.

As it stands, if we were to go through and implement every Linux system
call required by Go and build them into Akaros, in theory, we would have
a fully working system that mimiced the behaviour of Linux.  Obviously
we don't want to do that, but now we have a starting point for where to
dive in.