Fix up singal handling for akaros
I added a new defaulti sighandler function that all signals get
initalized to at start up (in runtime/parlib/signal.go).  This handler
basically follows the same path as the default fault handler for the
sigtable set up in the runtime.

Remember we have 2 paths for signals in akaros.  Thos ethat are user
generated (which only run through go code in parlib/signal.go), and
traps which are kernel generated, which go through the c runtime code.

The problem was that the default handler for these two setups differed.
I basically put the magic in place so that default handlers generated by
both paths run the same code when no new handler is installed on top of
them.