Move default CGO_FLAGS into build script
The original way of doin gthis used the "-pthread" options which was OK
so long as the pthread library didn't have any dependancies.  If
pthreads requires us to link another library to work, then we need to do
soemthing like what I'm doing now.  Possibly need to revisit this in the
future.