Added a bunch of support scripts for akaros
These scripts simplify the task of setting up and running a 9p server
for akaros, as well as running a qemu instance of akaros that is capable of
running simple rpc calls.  The rest of the go scripts take advantage of
this rpc capability to run go executables remotely, but report the
results locally.