Add parlib calls for chmod() and fchmod()
Normally I make direct syscalls rather than going through libc, but the
implementations for chmod and fchmod are now based on the new wstat
syscall, and are a bit more complicated than the other syscalls we've
implemented so far. It's just easier to make the direct libc calls and
not deal with all the convD2M stuff in go at all.