Fix TestExtraFiles (sort of)
Go runtime for akaros requires the alarm service to be open, which has 2
open files, one for the timerfd, and one for its ctl.  Don't close them!
Before these are open though, another file (which is subsequently
properly closed) comes along and claims fd 3, giving the alarm service
fds 4 and 5, so we need to bump n up by 3 instead of just 2.  Doh!