Up some delays and timeouts in the test
This doesn't affect what the test is testing, but makes it more likely
for the test to succeed on OSs that timeout on writes differently than
in a unix like system which uses epoll.

Specifically, Akaros times out each write using an alarm to trigger an
abort_syscall() system call.  This method causes problems for the test,
however, because system calls can only be aborted at specific
synchronization points, not while in the middle of an operation (i.e.
they can only be aborted when sitting inside a rendvous sleep).  If the
abortion fails, we reset the alarm with a certain delay (currently 1
second) and retry the abort later.

Because of the periodic nature of the abort retries and the fact that
the read and the write are mostly running in lock-step, its hard to
actually hit the case where we end up aborting the write.
Adding these new delays and the new timeout value compensate for this
and make the test more reliable on Akaros without affecting other
systems.