forkAndExec doesnt need a pipe on akaros
On akaros, we don't actually perform a fork/exec combination.  On
systems taht do, a pipe was necessary to pass the error code of the
forked process back to its parent in the case that execing inside it
failed for any reason.

On akaros, we don't need this pipe at all because of the way we do the
proc_create(), proc_run() combination. Previously we were just messing
things up by opening the pipe, closing the write side immediately upon
calling forkAndExecInChild, and then closing the read side twice. The
way we were doing things were remnants of a weird combiantion of doing
things the plan9 way, the unix way, and the akaros way without much
thought.  Hopefully it's doing the right thing now aftre a little more
scrutiny.